The 'lopensemble' package.

The lopensemble package provides an easy way to stack models based on the Continuous Ranked Probability Score (CRPS) based on predictive samples. Given some training data with true observed values as well as predictions generated from different models, lopensemble finds the optimal weights to form an ensemble from these models. Using these weights, lopensemble can then provide samples from the optimal model mixture by drawing from the predictice samples of those models in the correct proportion. This gives a true mixture model solely based on predictive samples and is in this regard superior to other ensembling techniques like Bayesian Model Averaging.

Currently stacking using CRPS is supported. Weights are generated using the crps_weights function. With these weights and predictive samples, mixture_from_samples can be used to obtain predictive samples from the optimal mixture model.

Author

Maintainer: Nikos Bosse nikosbosse@gmail.com (ORCID)

Authors:

Yuling Yao yy2619@columbia.edu

Sam Abbott contact@samabbott.co.uk (ORCID)

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID)